Make restored_layer_action public
