+++ buildpacks/ruby/src/layers/shared.rs
@@ -1,10 +1,9 @@
 use cache_diff::CacheDiff;
-use commons::display::SentenceList;
-use commons::layer::cache_buddy::invalid_metadata_action;
 pub(crate) use commons::layer::cache_buddy::Meta;
+use commons::layer::cache_buddy::{invalid_metadata_action, restored_layer_action};
 use libcnb::build::BuildContext;
 use libcnb::data::layer::LayerName;
-use libcnb::layer::{CachedLayerDefinition, LayerRef, RestoredLayerAction};
+use libcnb::layer::{CachedLayerDefinition, LayerRef};
 use magic_migrate::TryMigrate;
 use serde::ser::Serialize;
 use std::fmt::Debug;
@@ -33,29 +32,6 @@ where
     )?;
     layer_ref.write_metadata(metadata)?;
     Ok(layer_ref)
-}
-
-/// Standardizes formatting for layer cache clearing behavior
-///
-/// If the diff is empty, there are no changes and the layer is kept and the old data is returned
-/// If the diff is not empty, the layer is deleted and the changes are listed
-pub(crate) fn restored_layer_action<M>(old: &M, now: &M) -> (RestoredLayerAction, Meta<M>)
-where
-    M: CacheDiff + Clone,
-{
-    let diff = now.diff(old);
-    if diff.is_empty() {
-        (RestoredLayerAction::KeepLayer, Meta::Data(old.clone()))
-    } else {
-        (
-            RestoredLayerAction::DeleteLayer,
-            Meta::Message(format!(
-                "Clearing cache due to {changes}: {differences}",
-                changes = if diff.len() > 1 { "changes" } else { "change" },
-                differences = SentenceList::new(&diff)
-            )),
-        )
-    }
 }
 
 /// Removes ANSI control characters from a string
@@ -117,7 +93,7 @@ mod tests {
     use crate::RubyBuildpack;
     use core::panic;
     use libcnb::data::layer_name;
-    use libcnb::layer::{EmptyLayerCause, InvalidMetadataAction, LayerState};
+    use libcnb::layer::{EmptyLayerCause, InvalidMetadataAction, LayerState, RestoredLayerAction};
     use magic_migrate::{migrate_toml_chain, try_migrate_deserializer_chain, Migrate, TryMigrate};
     use serde::Deserializer;
     use std::convert::Infallible;
