+++ commons/Cargo.toml
@@ -29,6 +29,7 @@ walkdir = "2"
 filetime = "0.2"
 magic_migrate = "1.0"
 toml = "0.8"
+cache_diff = "1.0"
 
 [dev-dependencies]
 filetime = "0.2"
